criei um novo arquivo chamado routes, para dividir o código e deixar ele mais estruturado
